feat (auth-ms): agrega exceptions

+++ apps/dapp-auth-ms/src/modules/users/application/use-cases/delete-user.use-case.ts
@@ -1,6 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { UserRepositoryType } from '../../domain/repositories/user-repository.types';
 import { UserRepository } from '../../domain/repositories/user.repository.interface';
+import { UserNotFoundException } from '../exceptions/UserNotFoundException';
 
 @Injectable()
 export class DeleteUserUseCase {
@@ -10,6 +11,10 @@ export class DeleteUserUseCase {
   ) {}
 
   async execute(id: string) {
+    const user = await this.userRepository.findById(id);
+    if (!user) {
+      throw new UserNotFoundException();
+    }
     return this.userRepository.delete(id);
   }
 }
